requirements.txt
streamlit
streamlit-chat
pinecone-client
openai==0.28
python-dotenv
